Summaries refactor
