+++ src/summaries/overview/patient-chart-summary.component.tsx
@@ -1,13 +1,13 @@
 import React from "react";
 import { RouteComponentProps } from "react-router";
 import styles from "./patient-chart-summary.css";
-import HeightAndWeightBriefDetails from "../widgets/heightandweight/heightandweight-brief-details.component";
-import VitalsBriefSummary from "../widgets/vitals/vitals-brief-summary.component";
-import ConditionsCard from "../widgets/conditions/conditions-card.component";
-import AllergyCard from "../widgets/allergies/allergy-card.component";
-import NotesCard from "../widgets/notes/notes-card.component";
-import ProgramsCard from "../widgets/programs/programs-card.component";
-import MedicationsSummary from "../widgets/medications/medications-summary.component";
+import HeightAndWeightBriefDetails from "../../widgets/heightandweight/heightandweight-brief-details.component";
+import VitalsBriefSummary from "../../widgets/vitals/vitals-brief-summary.component";
+import ConditionsCard from "../../widgets/conditions/conditions-card.component";
+import AllergyCard from "../../widgets/allergies/allergy-card.component";
+import NotesCard from "../../widgets/notes/notes-card.component";
+import ProgramsCard from "../../widgets/programs/programs-card.component";
+import MedicationsSummary from "../../widgets/medications/medications-summary.component";
 
 export default function PatientChartSummary(props: PatientChartSummaryProps) {
   const config = [
