--- src/summary/summary-section-header.css
@@ -1,18 +1,0 @@
-a.headerLink:global(.omrs-unstyled) {
-  height: 3.75rem;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 0 0.25rem 0.25rem 1rem;
-  border-radius: 0.25rem;
-  background-color: var(--omrs-color-bg-lowest-contrast);
-  margin-top: 1rem;
-}
-
-:global(.omrs-breakpoint-lt-tablet) .headerLink {
-  margin-bottom: 0.5rem;
-}
-
-:global(.omrs-breakpoint-gt-phone) .headerLink {
-  margin-bottom: 1.25rem;
-}
