+++ translations/en.json
@@ -1,9 +1,4 @@
 {
-  "Active Medications": "Active Medications",
-  "Active Programs": "Active Programs",
   "Medications -current": "Medications -current",
-  "Medications-previous": "Medications-previous",
-  "Since": "Since",
-  "care programs": "Care Programs",
-  "conditions": "Conditions"
+  "Medications-previous": "Medications-previous"
 }
